fix: 修复多项强度控制与检测逻辑问题
