--- DG-LAB-VRCOSC-master/requirements.txt
@@ -1,13 +1,0 @@
-pydglab-ws
-qrcode
-python-osc
-colorlog
-pillow
-pyyaml
-psutil
-PySide6
-qasync
-websockets
-aiohttp
-requests
-zeroconf
