--- DG-LAB-VRCOSC-master/src/locales/en.yml
@@ -1,172 +1,0 @@
-# English translation file
-main:
-  title: "DG-Lab WebSocket Controller for VRChat"
-  tabs:
-    network: "Network Config"
-    controller: "Controller Settings"
-    osc: "OSC Parameters"
-    sps: "SPS Integration"
-    ton: "ToN Game Integration"
-    log: "Log Viewer"
-  settings:
-    language: "Language"
-
-network_tab:
-  title: "Network Configuration"
-  interface: "Network Interface"
-  ip_address: "IP Address"
-  websocket_port: "WebSocket Port"
-  osc_port: "OSC Port"
-  connect: "Connect"
-  disconnect: "Disconnect"
-  status: "Status"
-  online: "Online"
-  offline: "Offline, scan QR code after connecting"
-  start: "Start"
-  remote_address: "Remote Address"
-  enable_remote: "Remote Mode"
-  get_public_ip: "Get Public IP"
-  please_enter_valid_ip: "Please enter a valid IP address"
-
-controller_tab:
-  title: "Controller Settings"
-  current_channel: "Current Channel"
-  intensity: "Channel Intensity"
-  channel_intensity: "Channel Intensity"
-  intensity_limit: "Intensity Limit"
-  target_strength: "Target Strength"
-  mode: "Control Mode"
-  pulse_mode: "Pulse Mode"
-  waveform: "Waveform Mode"
-  pulse_width: "Pulse Width"
-  frequency: "Frequency"
-  enable_chatbox: "Enable ChatBox Status Display"
-  strength_step: "Fire Intensity Step"
-  adjust_step: "Adjust Intensity Step"
-  command_sources: "Command Source Control"
-  enable_gui_control: "Enable UI Control"
-  enable_soundpad: "Enable Soundpad Control"
-  current_panel: "Current Panel Channel"
-  not_set: "Not Set"
-  enable_interaction: "Enable Interaction Control"
-  interaction_control: "Channel Interaction"
-  enable_game_integration: "Enable Game Integration"
-
-osc_tab:
-  title: "OSC Parameters"
-  parameter_name: "Parameter Name"
-  parameter_type: "Parameter Type"
-  add: "Add"
-  remove: "Remove"
-  save: "Save"
-  cancel: "Cancel"
-  address_placeholder: "OSC Address"
-  channel_range_a: "Channel A Range"
-  channel_range_b: "Channel B Range"
-  min_value: "Min"
-  max_value: "Max"
-
-sps_tab:
-  title: "SPS Integration"
-  refresh: "Detect Current Avatar"
-  not_scanned: "No SPS zones detected yet"
-  scanning: "Reading VRChat OSCQuery..."
-  found: "Detected {count} SPS zones from {host}"
-  scan_failed: "Detection failed: {error}"
-  max_strength: "Max Strength"
-  channel_range_a: "Channel A Range"
-  channel_range_b: "Channel B Range"
-  bulk_sources: "Batch Sources"
-  bulk_source_my_parts: "My Plug/Socket"
-  sources: "Sources"
-  source_own_hands: "Own Hands"
-  source_other_hands: "Other Hands"
-  source_my_plugs: "My Plugs"
-  source_other_plugs: "Other Plugs"
-  source_other_sockets: "Other Sockets"
-  source_my_sockets: "My Sockets"
-
-ton_tab:
-  title: "ToN Game Integration"
-  enable: "Enable ToN Damage System"
-  test: "Test"
-  user_display_name: "User Display Name"
-  websocket_status: "WebSocket Status"
-  connected: "Connected"
-  disconnected: "Disconnected"
-  error: "Error"
-  accumulated_damage: "Accumulated Damage"
-  damage_reduction: "Damage Reduction per Second"
-  damage_strength: "Damage Strength Limit"
-  death_penalty: "Death Penalty Strength"
-  death_penalty_time: "Death Penalty Duration (sec)"
-  unknown: "Unknown"
-
-log_tab:
-  title: "Log Viewer"
-  clear: "Clear"
-  level: "Log Level"
-  save: "Save Log"
-  simple_log: "Simple Log"
-  debug_info: "Debug Information"
-  controller_params: "DGLabController Parameters"
-  loading_params: "Loading controller parameters..."
-  last_command_source: "Last Command Source"
-  last_command_time: "Last Command Time"
-  controller_status: "Controller Status"
-  device_online: "Device Online"
-  fire_strength: "Fire Strength"
-  chatbox_status: "ChatBox Status"
-  current_channel: "Current Channel"
-  command_status: "Command Type Status"
-  gui_commands: "GUI Commands"
-  panel_commands: "Panel Commands"
-  interaction_commands: "Interaction Commands"
-  channel_interaction: "Channel Interaction"
-  game_commands: "Game Integration Commands"
-  command_queue: "Command Queue Status"
-  queue_size: "Current Queue Size"
-  controller_not_initialized: "Controller not initialized"
-
-about_tab:
-  title: "About"
-  current_version: "Current Version"
-  check_update: "Check for Updates"
-  automatic_update_check: "Enable Automatic Update Check"
-  already_latest_version: "Already the latest version"
-  discover_new_version: "New version available"
-  new_version: "New Version"
-  update: "Update Content"
-  update_now: "Update Now"
-  cancel: "Cancel"
-  feedback: "Feedback"
-
-sps_tab:
-  title: "SPS Integration"
-  refresh: "Detect Current Avatar"
-  not_scanned: "No SPS zones detected yet"
-  scanning: "Reading VRChat OSCQuery..."
-  found: "Detected {count} SPS zones from {host}"
-  scan_failed: "Detection failed: {error}"
-  max_strength: "Max Strength"
-  channel_range_a: "Channel A Range"
-  channel_range_b: "Channel B Range"
-  bulk_sources: "Batch Sources"
-  bulk_source_my_parts: "My Plug/Socket"
-  sources: "Sources"
-  source_own_hands: "Own Hands"
-  source_other_hands: "Other Hands"
-  source_my_plugs: "My Plugs"
-  source_other_plugs: "Other Plugs"
-  source_other_sockets: "Other Sockets"
-  source_my_sockets: "My Sockets"
-  progress_starting: "Preparing to scan..."
-  progress_connecting: "Connecting to VRChat OSCQuery..."
-  progress_fetching_nodes: "Fetching parameter tree..."
-  progress_parsing_avatar: "Parsing current Avatar..."
-  progress_scanning_zones: "Scanning OGB/SPS zones..."
-  progress_merging: "Merging bindings..."
-  progress_updating_ui: "Updating zone list..."
-  progress_saving: "Saving configuration..."
-  progress_syncing_osc: "Syncing to OSC Parameters tab..."
-  progress_done: "Scan complete!"
