--- DG-LAB-VRCOSC-master/src/locales/ja.yml
@@ -1,142 +1,0 @@
-# 日本語翻訳ファイル
-main:
-  title: "DG-Lab WebSocket Controller for VRChat"
-  tabs:
-    network: "ネットワーク設定"
-    controller: "コントローラー設定"
-    osc: "OSCパラメータ"
-    sps: "SPS連動"
-    ton: "ToNゲーム連携"
-    log: "ログビューア"
-  settings:
-    language: "言語"
-
-network_tab:
-  title: "ネットワーク設定"
-  interface: "ネットワークインターフェース"
-  ip_address: "IPアドレス"
-  websocket_port: "WebSocketポート"
-  osc_port: "OSCポート"
-  connect: "接続"
-  disconnect: "切断"
-  status: "状態"
-  online: "オンライン"
-  offline: "オフライン、接続後にQRコードをスキャンしてください"
-  start: "開始"
-  remote_address: "リモートアドレス"
-  enable_remote: "遠隔モードを開始する"
-  get_public_ip: "パブリックIPを取得する"
-  please_enter_valid_ip: "有効なIPを入力してください"
-
-controller_tab:
-  title: "コントローラー設定"
-  current_channel: "現在のチャンネル"
-  intensity: "チャンネル強度"
-  channel_intensity: "チャンネル強度"
-  intensity_limit: "強度上限"
-  target_strength: "目標強度"
-  mode: "制御モード"
-  pulse_mode: "パルスモード"
-  waveform: "波形モード"
-  pulse_width: "パルス幅"
-  frequency: "周波数"
-  enable_chatbox: "ChatBoxステータス表示を有効化"
-  strength_step: "発火強度ステップ"
-  adjust_step: "調整強度ステップ"
-  command_sources: "コマンドソース制御"
-  enable_gui_control: "UI制御を有効化"
-  enable_soundpad: "Soundpad制御を有効化"
-  current_panel: "現在のパネルチャンネル"
-  not_set: "未設定"
-  enable_interaction: "インタラクション制御を有効化"
-  interaction_control: "チャンネルインタラクション"
-  enable_game_integration: "ゲーム連携を有効化"
-
-osc_tab:
-  title: "OSCパラメータ"
-  parameter_name: "パラメータ名"
-  parameter_type: "パラメータタイプ"
-  add: "追加"
-  remove: "削除"
-  save: "保存"
-  cancel: "キャンセル"
-  address_placeholder: "OSCアドレス"
-  channel_range_a: "チャンネルAの範囲"
-  channel_range_b: "チャンネルBの範囲"
-  min_value: "最小"
-  max_value: "最大"
-
-sps_tab:
-  title: "SPS連動"
-  refresh: "現在のAvatarを検出"
-  not_scanned: "SPSゾーンはまだ検出されていません"
-  scanning: "VRChat OSCQueryを読み込み中..."
-  found: "{host} から {count} 個のSPSゾーンを検出しました"
-  scan_failed: "検出失敗: {error}"
-  max_strength: "最大強度"
-  channel_range_a: "チャンネルAの範囲"
-  channel_range_b: "チャンネルBの範囲"
-  bulk_sources: "入力元の一括切替"
-  bulk_source_my_parts: "自分のPlug/Socket"
-  sources: "入力元"
-  source_own_hands: "自分の手"
-  source_other_hands: "相手の手"
-  source_my_plugs: "自分のPlug"
-  source_other_plugs: "相手のPlug"
-  source_other_sockets: "相手のSocket"
-  source_my_sockets: "自分のSocket"
-
-ton_tab:
-  title: "ToNゲーム連携"
-  enable: "ToNダメージシステムを有効化"
-  test: "テスト"
-  user_display_name: "ユーザー表示名"
-  websocket_status: "WebSocket状態"
-  connected: "接続済み"
-  disconnected: "未接続"
-  error: "エラー"
-  accumulated_damage: "累積ダメージ"
-  damage_reduction: "1秒あたりのダメージ軽減"
-  damage_strength: "ダメージ強度上限"
-  death_penalty: "死亡ペナルティ強度"
-  death_penalty_time: "死亡ペナルティ持続時間（秒）"
-  unknown: "不明"
-
-log_tab:
-  title: "ログビューア"
-  clear: "クリア"
-  level: "ログレベル"
-  save: "ログを保存"
-  simple_log: "シンプルログ"
-  debug_info: "デバッグ情報"
-  controller_params: "DGLabControllerパラメータ"
-  loading_params: "コントローラーパラメータを読み込み中..."
-  last_command_source: "最終コマンドソース"
-  last_command_time: "最終コマンド時刻"
-  controller_status: "コントローラー状態"
-  device_online: "デバイスオンライン"
-  fire_strength: "発火強度"
-  chatbox_status: "ChatBox状態"
-  current_channel: "現在のチャンネル"
-  command_status: "コマンドタイプ状態"
-  gui_commands: "GUIコマンド"
-  panel_commands: "パネルコマンド"
-  interaction_commands: "インタラクションコマンド"
-  channel_interaction: "チャンネルインタラクション"
-  game_commands: "ゲーム連携コマンド"
-  command_queue: "コマンドキュー状態"
-  queue_size: "現在のキューサイズ"
-  controller_not_initialized: "コントローラーが初期化されていません" 
-
-about_tab:
-  title: "アプリについて"
-  current_version: "現在のバージョン"
-  check_update: "更新を確認"
-  automatic_update_check: "自動更新確認を有効にする"
-  already_latest_version: "現在のバージョンは最新です"
-  discover_new_version: "新しいバージョンが見つかりました"
-  new_version: "新バージョン"
-  update: "更新内容"
-  update_now: "今すぐ更新"
-  cancel: "キャンセル" 
-  feedback: "フィードバック"
