--- DG-LAB-VRCOSC-master/src/locales/zh.yml
@@ -1,172 +1,0 @@
-# 中文翻译文件
-main:
-  title: "DG-Lab WebSocket Controller for VRChat"
-  tabs:
-    network: "网络配置"
-    controller: "控制器设置"
-    osc: "OSC参数配置"
-    sps: "SPS联动"
-    ton: "ToN游戏联动"
-    log: "日志查看"
-  settings:
-    language: "语言"
-
-network_tab:
-  title: "网络配置"
-  interface: "网卡选择"
-  ip_address: "IP地址"
-  websocket_port: "WebSocket端口"
-  osc_port: "OSC端口"
-  connect: "连接"
-  disconnect: "断开"
-  status: "状态"
-  online: "在线"
-  offline: "未连接，请点击连接后扫描二维码"
-  start: "启动"
-  remote_address: "远程地址"
-  enable_remote: "开启异地模式"
-  get_public_ip: "获取公网IP"
-  please_enter_valid_ip: "请输入有效的IP地址"
-
-controller_tab:
-  title: "控制器设置"
-  current_channel: "当前通道"
-  intensity: "通道强度"
-  channel_intensity: "通道强度"
-  intensity_limit: "强度上限"
-  target_strength: "目标强度"
-  mode: "控制模式"
-  pulse_mode: "脉冲模式"
-  waveform: "波形模式"
-  pulse_width: "脉宽"
-  frequency: "频率"
-  enable_chatbox: "启用ChatBox状态显示"
-  strength_step: "开火强度步进"
-  adjust_step: "调节强度步进"
-  command_sources: "命令来源控制"
-  enable_gui_control: "启用程序界面控制"
-  enable_soundpad: "启用Soundpad控制"
-  current_panel: "面板当前控制通道"
-  not_set: "未设置"
-  enable_interaction: "启用交互方式控制"
-  interaction_control: "通道交互控制"
-  enable_game_integration: "启用游戏联动控制"
-
-osc_tab:
-  title: "OSC参数配置"
-  parameter_name: "参数名称"
-  parameter_type: "参数类型"
-  add: "添加"
-  remove: "删除"
-  save: "保存"
-  cancel: "取消"
-  address_placeholder: "OSC 地址"
-  channel_range_a: "A通道范围"
-  channel_range_b: "B通道范围"
-  min_value: "最小"
-  max_value: "最大"
-
-sps_tab:
-  title: "SPS联动"
-  refresh: "自动探测当前Avatar"
-  not_scanned: "尚未探测到SPS区域"
-  scanning: "正在读取VRChat OSCQuery..."
-  found: "已从 {host} 探测到 {count} 个SPS区域"
-  scan_failed: "探测失败: {error}"
-  max_strength: "最高强度"
-  channel_range_a: "A通道范围"
-  channel_range_b: "B通道范围"
-  bulk_sources: "批量来源"
-  bulk_source_my_parts: "自己的Plug/Socket"
-  sources: "来源"
-  source_own_hands: "自己的手"
-  source_other_hands: "别人的手"
-  source_my_plugs: "自己的Plug"
-  source_other_plugs: "别人的Plug"
-  source_other_sockets: "别人的Socket"
-  source_my_sockets: "自己的Socket"
-
-ton_tab:
-  title: "ToN游戏联动"
-  enable: "启用ToN伤害系统"
-  test: "测试"
-  user_display_name: "用户显示名称"
-  websocket_status: "WebSocket状态"
-  connected: "已连接"
-  disconnected: "未连接"
-  error: "错误"
-  accumulated_damage: "累计伤害"
-  damage_reduction: "每秒伤害减免强度"
-  damage_strength: "伤害对应强度上限"
-  death_penalty: "死亡惩罚强度"
-  death_penalty_time: "死亡惩罚持续时间 (秒)"
-  unknown: "未知"
-
-log_tab:
-  title: "日志查看"
-  clear: "清除"
-  level: "日志级别"
-  save: "保存日志"
-  simple_log: "简约日志"
-  debug_info: "调试信息"
-  controller_params: "DGLabController参数"
-  loading_params: "正在加载控制器参数..."
-  last_command_source: "最近命令来源"
-  last_command_time: "最近命令时间"
-  controller_status: "控制器状态"
-  device_online: "设备在线"
-  fire_strength: "一键开火强度"
-  chatbox_status: "ChatBox状态"
-  current_channel: "当前选择通道"
-  command_status: "命令类型状态"
-  gui_commands: "GUI命令"
-  panel_commands: "面板命令"
-  interaction_commands: "交互命令"
-  channel_interaction: "通道交互"
-  game_commands: "游戏联动命令"
-  command_queue: "命令队列状态"
-  queue_size: "当前队列大小"
-  controller_not_initialized: "控制器未初始化"
-
-about_tab:
-  title: "关于"
-  current_version: "当前版本"
-  check_update: "检查更新"
-  automatic_update_check: "启用自动更新检查"
-  already_latest_version: "当前已是最新版本"
-  discover_new_version: "发现新版本"
-  new_version: "新版本"
-  update: "更新内容"
-  update_now: "立即更新"
-  cancel: "取消"
-  feedback: "问题反馈"
-
-sps_tab:
-  title: "SPS联动"
-  refresh: "自动探测当前Avatar"
-  not_scanned: "尚未探测到SPS区域"
-  scanning: "正在读取VRChat OSCQuery..."
-  found: "已从 {host} 探测到 {count} 个SPS区域"
-  scan_failed: "探测失败: {error}"
-  max_strength: "最大强度"
-  channel_range_a: "A通道范围"
-  channel_range_b: "B通道范围"
-  bulk_sources: "批量来源"
-  bulk_source_my_parts: "我的插槽/插头"
-  sources: "来源"
-  source_own_hands: "自己的手"
-  source_other_hands: "别人的手"
-  source_my_plugs: "我的插头"
-  source_other_plugs: "别人的插头"
-  source_other_sockets: "别人的插槽"
-  source_my_sockets: "我的插槽"
-  progress_starting: "正在准备扫描..."
-  progress_connecting: "正在连接 VRChat OSCQuery 服务..."
-  progress_fetching_nodes: "正在读取参数树..."
-  progress_parsing_avatar: "正在解析当前 Avatar..."
-  progress_scanning_zones: "正在扫描 OGB/SPS 区域..."
-  progress_merging: "正在合并绑定配置..."
-  progress_updating_ui: "正在更新界面列表..."
-  progress_saving: "正在保存配置..."
-  progress_syncing_osc: "正在同步到 OSC 参数标签页..."
-  progress_done: "扫描完成！"
