--- DG-LAB-VRCOSC-master/src/version.py
@@ -1,1 +1,0 @@
-VERSION = "v0.4.3"
