+++ src/locales/en.yml
@@ -140,3 +140,33 @@ about_tab:
   update_now: "Update Now"
   cancel: "Cancel"
   feedback: "Feedback"
+
+sps_tab:
+  title: "SPS Integration"
+  refresh: "Detect Current Avatar"
+  not_scanned: "No SPS zones detected yet"
+  scanning: "Reading VRChat OSCQuery..."
+  found: "Detected {count} SPS zones from {host}"
+  scan_failed: "Detection failed: {error}"
+  max_strength: "Max Strength"
+  channel_range_a: "Channel A Range"
+  channel_range_b: "Channel B Range"
+  bulk_sources: "Batch Sources"
+  bulk_source_my_parts: "My Plug/Socket"
+  sources: "Sources"
+  source_own_hands: "Own Hands"
+  source_other_hands: "Other Hands"
+  source_my_plugs: "My Plugs"
+  source_other_plugs: "Other Plugs"
+  source_other_sockets: "Other Sockets"
+  source_my_sockets: "My Sockets"
+  progress_starting: "Preparing to scan..."
+  progress_connecting: "Connecting to VRChat OSCQuery..."
+  progress_fetching_nodes: "Fetching parameter tree..."
+  progress_parsing_avatar: "Parsing current Avatar..."
+  progress_scanning_zones: "Scanning OGB/SPS zones..."
+  progress_merging: "Merging bindings..."
+  progress_updating_ui: "Updating zone list..."
+  progress_saving: "Saving configuration..."
+  progress_syncing_osc: "Syncing to OSC Parameters tab..."
+  progress_done: "Scan complete!"
