+++ src/locales/zh.yml
@@ -140,3 +140,33 @@ about_tab:
   update_now: "立即更新"
   cancel: "取消"
   feedback: "问题反馈"
+
+sps_tab:
+  title: "SPS联动"
+  refresh: "自动探测当前Avatar"
+  not_scanned: "尚未探测到SPS区域"
+  scanning: "正在读取VRChat OSCQuery..."
+  found: "已从 {host} 探测到 {count} 个SPS区域"
+  scan_failed: "探测失败: {error}"
+  max_strength: "最大强度"
+  channel_range_a: "A通道范围"
+  channel_range_b: "B通道范围"
+  bulk_sources: "批量来源"
+  bulk_source_my_parts: "我的插槽/插头"
+  sources: "来源"
+  source_own_hands: "自己的手"
+  source_other_hands: "别人的手"
+  source_my_plugs: "我的插头"
+  source_other_plugs: "别人的插头"
+  source_other_sockets: "别人的插槽"
+  source_my_sockets: "我的插槽"
+  progress_starting: "正在准备扫描..."
+  progress_connecting: "正在连接 VRChat OSCQuery 服务..."
+  progress_fetching_nodes: "正在读取参数树..."
+  progress_parsing_avatar: "正在解析当前 Avatar..."
+  progress_scanning_zones: "正在扫描 OGB/SPS 区域..."
+  progress_merging: "正在合并绑定配置..."
+  progress_updating_ui: "正在更新界面列表..."
+  progress_saving: "正在保存配置..."
+  progress_syncing_osc: "正在同步到 OSC 参数标签页..."
+  progress_done: "扫描完成！"
